--- singleGUIAutomationsProject/src/test/java/PageObjects/RewardsPage.java
@@ -1,4 +1,0 @@
-package PageObjects;
-
-public class RewardsPage {
-}
