--- singleGUIAutomationsProject/src/test/java/Runner/IndividualCoverRunner.java
@@ -1,14 +1,0 @@
-package Runner;
-
-import io.cucumber.junit.CucumberOptions;
-import net.serenitybdd.cucumber.CucumberWithSerenity;
-import org.junit.runner.RunWith;
-
-@RunWith(CucumberWithSerenity.class)
-@CucumberOptions(
-        features = "src/test/resources/features/IndividualCover.feature",
-        glue = "StepDefinitions"
-
-)
-public class IndividualCoverRunner {
-}
