+++ src/test/java/PageObjects/RewardsPage.java
@@ -1,0 +1,4 @@
+package PageObjects;
+
+public class RewardsPage {
+}
