+++ src/test/java/PageObjects/startSalePage.java
@@ -1,0 +1,17 @@
+package PageObjects;
+
+import net.serenitybdd.core.pages.PageObject;
+import net.serenitybdd.core.pages.PageObjects;
+import org.openqa.selenium.WebDriver;
+
+public class startSalePage extends PageObjects {
+
+    String StartSaleXpath = "//*[@id=bootstrap-table]/tbody/tr/td[9]/button[1]";
+    String SilverRewardsCheckBoxXpath = "//*[@id=\"bootstrap-table\"]/tbody/tr[1]/td[1]/input";
+    String BlueRewardsCheckBoxXpath = "//*[@id=\"bootstrap-table\"]/tbody/tr[2]/td[1]/input";
+
+
+    public startSalePage(WebDriver driver) {
+        super(driver);
+    }
+}
