+++ src/test/java/Runner/SingleGUITest.java
@@ -1,0 +1,18 @@
+package Runner;
+
+import io.cucumber.junit.CucumberOptions;
+import net.serenitybdd.cucumber.CucumberWithSerenity;
+import org.junit.runner.RunWith;
+
+
+@RunWith(CucumberWithSerenity.class)
+@CucumberOptions(
+        features = "src/test/resources/features/search/SingleGUITest.feature",
+        glue = "StepDefinitions"
+
+
+)
+public class SingleGUITest {
+
+
+}
