+++ src/test/java/SingleGUITest.java
@@ -1,0 +1,3 @@
+public class SingleGUITest {
+
+}
