--- src/test/java/starter/navigation/NavigateTo.java
@@ -1,12 +1,0 @@
-package starter.navigation;
-
-import net.serenitybdd.screenplay.Performable;
-import net.serenitybdd.screenplay.Task;
-import net.serenitybdd.screenplay.actions.Open;
-
-public class NavigateTo {
-    public static Performable theSearchHomePage() {
-        return Task.where("{0} opens the DuckDuckGo home page",
-                Open.browserOn().the(DuckDuckGoHomePage.class));
-    }
-}
