--- src/test/java/starter/search/SearchForm.java
@@ -1,8 +1,0 @@
-package starter.search;
-
-import net.serenitybdd.screenplay.targets.Target;
-
-class SearchForm {
-    static Target SEARCH_FIELD = Target.the("search field").locatedBy("#searchbox_input");
-
-}
