--- src/test/java/starter/stepdefinitions/ParameterDefinitions.java
@@ -1,20 +1,0 @@
-package starter.stepdefinitions;
-
-import io.cucumber.java.Before;
-import io.cucumber.java.ParameterType;
-import net.serenitybdd.screenplay.Actor;
-import net.serenitybdd.screenplay.actors.OnStage;
-import net.serenitybdd.screenplay.actors.OnlineCast;
-
-public class ParameterDefinitions {
-
-    @ParameterType(".*")
-    public Actor actor(String actorName) {
-        return OnStage.theActorCalled(actorName);
-    }
-
-    @Before
-    public void setTheStage() {
-        OnStage.setTheStage(new OnlineCast());
-    }
-}
