--- src/test/resources/features/search/search_by_keyword.feature
@@ -1,13 +1,0 @@
-Feature: Search by keyword
-
-  @green
-  Scenario: Searching for 'green'
-    Given Sergey is researching things on the internet
-    When he looks up "green"
-    Then he should see information about "green"
-
-  @red
-  Scenario: Searching for 'red'
-    Given Sergey is researching things on the internet
-    When he looks up "red"
-    Then he should see information about "red"
